requirements.txt
altgraph==0.17.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
autopep8==1.6.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
black==21.10b0
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
construct==2.10.67
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
executing==0.10.0
fasteners==0.14.1
fastjsonschema==2.16.1
filelock==3.6.0
Flask==2.1.3
future==0.18.2
glib==1.0.0
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.4
ifaddr==0.1.7
importlib-metadata==4.6.4
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
itsdangerous==2.1.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.13.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==0.8.4
mkchromecast==0.3.9
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mypy-extensions==0.4.3
natsort==8.0.2
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.11.0
notebook==6.4.12
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.12.4
proxy-tools==0.1.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pwquality==1.4.4
PyAutoGUI==0.9.53
pycairo==1.20.1
PyChromecast==9.4.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
PyGetWindow==0.0.9
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PyJWT==2.3.0
pykeepass==4.0.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.1
PyScreeze==0.1.28
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.1
pywebview==3.6.3
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
reportlab==3.6.8
requests==2.28.1
SecretStorage==3.3.1
Send2Trash==1.8.0
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.4.0
systemd-python==234
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.2
tornado==6.2
traitlets==5.3.0
typing-extensions==3.10.0.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.6
urllib3==1.26.13
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==4.0.2
wxPython==4.0.7
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zeroconf==0.38.3
zipp==1.0.0
